+++ ios/Classes/FLTBetterPlayerPlugin.m
@@ -186,7 +186,6 @@ AVPictureInPictureController *_pipController;
 }
 
 - (void)itemDidPlayToEndTime:(NSNotification*)notification {
-    
     if (_isLooping) {
         AVPlayerItem* p = [notification object];
         [p seekToTime:kCMTimeZero completionHandler:nil];
@@ -463,7 +462,11 @@ static inline CGFloat radiansToDegrees(CGFloat radians) {
     }
     
     if (_isPlaying) {
-        [_player play];
+        if (@available(iOS 10.0, *)) {
+            [_player playImmediatelyAtRate:1.0];
+        } else {
+            [_player play];
+        }
     } else {
         [_player pause];
     }
@@ -704,6 +707,26 @@ static inline CGFloat radiansToDegrees(CGFloat radians) {
 - (void)pictureInPictureController:(AVPictureInPictureController *)pictureInPictureController restoreUserInterfaceForPictureInPictureStopWithCompletionHandler:(void (^)(BOOL))completionHandler {
     [self setRestoreUserInterfaceForPIPStopCompletionHandler: true];
 }
+
+- (void) setAudioTrack:(NSString*) name index:(int) index{
+    AVMediaSelectionGroup *audioSelectionGroup = [[[_player currentItem] asset] mediaSelectionGroupForMediaCharacteristic: AVMediaCharacteristicAudible];
+    NSArray* options = audioSelectionGroup.options;
+    
+    
+    for (int index = 0; index < [options count]; index++) {
+        AVMediaSelectionOption* option = [options objectAtIndex:index];
+        NSArray *metaDatas = [AVMetadataItem metadataItemsFromArray:option.commonMetadata withKey:@"title" keySpace:@"comn"];
+        if (metaDatas.count > 0) {
+            NSString *title = ((AVMetadataItem*)[metaDatas objectAtIndex:0]).stringValue;
+            if (title == name && index == index ){
+                [[_player currentItem] selectMediaOption:option inMediaSelectionGroup: audioSelectionGroup];
+            }
+        }
+        
+    }
+}
+
+
 #endif
 // This workaround if you will change dataSource. Flutter engine caches CVPixelBufferRef and if you
 // return NULL from method copyPixelBuffer Flutter will use cached CVPixelBufferRef. If you will
@@ -1161,6 +1184,10 @@ NSMutableDictionary*  _artworkImageDict;
         } else if ([@"disablePictureInPicture" isEqualToString:call.method]){
             [player disablePictureInPicture];
             [player setPictureInPicture:false];
+        } else if ([@"setAudioTrack" isEqualToString:call.method]){
+            NSString* name = argsMap[@"name"];
+            int index = [argsMap[@"index"] intValue];
+            [player setAudioTrack:name index: index];
         }
         
         else {