+++ lib/src/core/better_player.dart
@@ -172,7 +172,7 @@ class BetterPlayerState extends State<BetterPlayer>
       Animation<double> animation,
       BetterPlayerControllerProvider controllerProvider) {
     return Scaffold(
-      resizeToAvoidBottomPadding: false,
+      resizeToAvoidBottomInset: false,
       body: Container(
         alignment: Alignment.center,
         color: Colors.black,