+++ lib/src/hls/better_player_hls_audio_track.dart
@@ -1,0 +1,16 @@
+///Representation of HLS audio track
+class BetterPlayerHlsAudioTrack {
+  ///Id of track inside HLS playlist
+  final int id;
+
+  ///Description of the audio
+  final String label;
+
+  ///Language code
+  final String language;
+
+  ///Url of audio track
+  final String url;
+
+  BetterPlayerHlsAudioTrack({this.id, this.label, this.language, this.url});
+}