+++ lib/src/hls/better_player_hls_subtitle.dart
@@ -1,7 +1,15 @@
+///Representation of HLS subtitle element.
 class BetterPlayerHlsSubtitle {
+  ///Language of the subtitle
   final String language;
+
+  ///Name of the subtitle
   final String name;
+
+  ///Url of the subtitle (master playlist)
   final String url;
+
+  ///Urls of specific files
   final List<String> realUrls;
 
   BetterPlayerHlsSubtitle({