+++ lib/src/video_player/video_player_platform_interface.dart
@@ -148,6 +148,10 @@ abstract class VideoPlayerPlatform {
         'isPictureInPictureEnabled() has not been implemented.');
   }
 
+  Future<void> setAudioTrack(int textureId, String name, int index) {
+    throw UnimplementedError('setAudio() has not been implemented.');
+  }
+
   /// Returns a widget displaying the video with a given textureID.
   Widget buildView(int textureId) {
     throw UnimplementedError('buildView() has not been implemented.');